Add files via upload
Hi Anton,

This file can be shown as an demo for our final product, which does all the calculation, but just in small scale. If I can finish the VBA part, it can be extended into unlimited size. Please help to make the print out table looks better, while I will be keeping working on the VBA part. The data is on tab "Solution data". Thx