New Interfaces 0318
I will use this interface to develop the professor assignment part. Please use this one as well for the course scheduling. 
The only fixed part in this file include "InputCourse" (column A:H), and "InputProf" (Column A:J). The other parts are still WIP and therefore can be changed if needed, especially the time_preference from faculties ("InputProf" column K: S). I'm still thinking about how to apply that.
Thx.