Interface_0322
Please take a look and let me know if you want to change anything. Please do not add any calculation columns in these tabs.
The green tabs are input tables for user and purple ones are output reports for users in the case they want to check the details about the report. 
Oblock tab block view of our result to review what kind of sessions will be in that block.
Ocourse tab is a course view of our result to show user each session will be taught by whom and at what time.
Oprof tab is a professor view of our result to show which professor will teach what sessions at what time.
Printout tab is a schedule grid as usual.
With these reports, I think it will be easier for user to make decisions and check the integrity of our result or make changes.

Please let me know your comments. Thx